+++ .github/workflows/publish.yaml
@@ -25,6 +25,7 @@ jobs:
 
       - run: (cd wasm-example/ && wasm-pack build -t web)
 
-      - uses: katyo/publish-crates@v1
-        with:
-          registry-token: ${{ secrets.CARGO_REGISTRY_TOKEN }}
+      - run: git checkout Cargo.lock
+
+      - name: Publish
+        run: cargo publish --token ${{ secrets.CARGO_REGISTRY_TOKEN }} --allow-dirty
