+++ Cargo.toml
@@ -1,7 +1,7 @@
 
 [package]
 name = "cirru_parser"
-version = "0.2.8"
+version = "0.2.10"
 authors = ["jiyinyiyong <jiyinyiyong@gmail.com>"]
 edition = "2024"
 license = "MIT"
