+++ src/parser.rs
@@ -25,6 +25,7 @@ find more on <http://text.cirru.org/> .
 
 mod error;
 mod primes;
+mod focus;
 mod s_expr;
 mod tree;
 mod writer;
@@ -46,7 +47,8 @@ use tree::{resolve_comma, resolve_dollar};
 
 pub use primes::{Cirru, CirruLexItem, CirruLexItemList, escape_cirru_leaf};
 pub use s_expr::format_to_lisp;
-pub use writer::{CirruOneLinerExt, CirruWriterOptions, focus_cirru_preview, format, format_expr_one_liner, generate_leaf};
+pub use focus::focus_cirru_preview;
+pub use writer::{CirruOneLinerExt, CirruWriterOptions, format, format_expr_one_liner, generate_leaf};
 
 /// Helper function to format and print a detailed error
 pub fn print_error(error: &CirruError, source_code: Option<&str>) {
